refactor: 아이콘 컴포넌트 index 파일 삭제
- 트리 쉐이킹이 되지 않아 번들 사이즈가 늘어나게 되므로 개별 import 방식 사용
