--- src/components/Icon/index.ts
@@ -1,5 +1,0 @@
-export { default as AlertCircleLine } from './AlertCircleLine';
-export { default as IconCheckCircle } from './IconCheckCircle';
-export { default as IconCheckLine } from './IconCheckLine';
-export { default as IconXLine } from './IconXLine';
-export { default as IconReset } from './IconReset';
